# netty: use custom Http2Headers class for encoding Metadata