Before:
```
Benchmark                              (headerCount)    Mode     Cnt     Score    Error  Units
HeadersBenchmark.convertClientHeaders             10  sample  119785   684.202 ±  4.531  ns/op
HeadersBenchmark.convertClientHeaders             20  sample  137935  1176.324 ± 34.101  ns/op
HeadersBenchmark.convertClientHeaders             50  sample  114824  2793.422 ± 76.779  ns/op
HeadersBenchmark.convertClientHeaders            100  sample  110157  5720.402 ± 65.462  ns/op
HeadersBenchmark.convertServerHeaders             10  sample  127963   640.391 ±  3.215  ns/op
HeadersBenchmark.convertServerHeaders             20  sample  134060  1190.960 ±  4.222  ns/op
HeadersBenchmark.convertServerHeaders             50  sample  116292  2713.037 ±  7.151  ns/op
HeadersBenchmark.convertServerHeaders            100  sample  115921  5427.773 ± 50.070  ns/op
```

After:
```
Benchmark                              (headerCount)    Mode     Cnt     Score    Error  Units
HeadersBenchmark.convertClientHeaders             10  sample  164584   520.479 ± 28.973  ns/op
HeadersBenchmark.convertClientHeaders             20  sample  191581   853.866 ± 25.689  ns/op
HeadersBenchmark.convertClientHeaders             50  sample  166162  1904.264 ± 30.057  ns/op
HeadersBenchmark.convertClientHeaders            100  sample  162615  3836.205 ±  8.345  ns/op
HeadersBenchmark.convertServerHeaders             10  sample   99981   418.113 ±  3.090  ns/op
HeadersBenchmark.convertServerHeaders             20  sample  198148   822.495 ± 23.354  ns/op
HeadersBenchmark.convertServerHeaders             50  sample  165877  1906.003 ±  5.161  ns/op
HeadersBenchmark.convertServerHeaders            100  sample  163624  3843.244 ± 45.640  ns/op
```

+cc: @ejona86