@carl-mastrangelo 2016-05-24
Tests are failing because of improperly used mocks.