@carl-mastrangelo 2016-05-24
Updated benchmarks improved by a lot for smaller header counts:

```
Benchmark                              (headerCount)    Mode     Cnt     Score     Error  Units
HeadersBenchmark.convertClientHeaders             10  sample  192945   436.981 ±  20.288  ns/op
HeadersBenchmark.convertClientHeaders             20  sample  189504   886.491 ±  22.591  ns/op
HeadersBenchmark.convertClientHeaders             50  sample  165510  1916.513 ±  36.414  ns/op
HeadersBenchmark.convertClientHeaders            100  sample  164783  3851.935 ± 230.092  ns/op
HeadersBenchmark.convertServerHeaders             10  sample  194939   427.932 ±   2.058  ns/op
HeadersBenchmark.convertServerHeaders             20  sample  197722   823.132 ±   3.068  ns/op
HeadersBenchmark.convertServerHeaders             50  sample  165089  1922.214 ±  30.240  ns/op
HeadersBenchmark.convertServerHeaders            100  sample  163625  3837.431 ±  28.306  ns/op
```