@buchgr 2016-05-25
@carl-mastrangelo great idea! I really like it. Not sure if you tested, but I assume the memory savings are HUGE? For example, can test actual memory usage of `DefaultHttp2Headers` and the new `GrpcHttp2Headers` with http://igstan.ro/posts/2014-09-23-calculating-an-object-graphs-size-on-the-jvm.html

 I feel stupid. Doing it that way from the beginning, we could have saved a lot of work and headaches trying to fix the Netty HTTP/2 header @ejona86 @nmittler @louiscryan @Scottmitch 😅