@Scottmitch 2016-05-25
@carl-mastrangelo - Very nice! It would be even nicer if we could have something similar to this upstream :). Something like `ReadOnlyHttp2Headers` (not modified after construction)? The "getter" methods which are currently not supported could just be implemented as iterative searches to start. We could punt on the "getter type conversion methods" initially too. Any interest in submitting a PR for this upstream?