@carl-mastrangelo 2016-05-25
@Scottmitch yeah actually.  These headers are currently idealized for gRPC specifically (notice the size difference between GrpcHttp2Headers and AbstractHttp2Headers).  This class is playing fast and loose with Iteration order since it is making assumptions about how the encoder will access it.  

I don't know how general purpose these would need to become in order to be useful to netty.  If these end up working out, and don't get reverted within a few days, I'll put together a PR.