@Scottmitch 2016-05-25
> if these end up working out, and don't get reverted within a few days

Such confidence :) I'm looking forward another PR upstream 👍