@carl-mastrangelo 2016-05-25
@buchgr Excellent idea.  Allocation is down from about 140 to 40!  We have a custom allocation recorder that lists every single allocation  Right now the biggest consumer is Metadata:

Code:
```
io.grpc.netty.Utils.convertClientHeaders(metadata, scheme, defaultPath, authority, userAgent);
```

```
I just allocated the object [[B@732f29af of type [B whose size is 96
It's an array of size 20
I just allocated the object [key-0=[glasgo], key-1=[glasgo], key-2=[glasgo], key-3=[glasgo], key-4=[glasgo], key-5=[glasgo], key-6=[glasgo], key-7=[glasgo], key-8=[glasgo], key-9=[glasgo]] of type java/util/LinkedHashMap$LinkedEntrySet whose size is 16
I just allocated the object java.util.LinkedHashMap$LinkedEntryIterator@d3957fe of type java/util/LinkedHashMap$LinkedEntryIterator whose size is 32
I just allocated the object java.util.ArrayList$Itr@6622fc65 of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@299321e2 of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@23fb172e of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@64ba3208 of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@158a3b2e of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@25a6944c of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@5e1fa5b1 of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@6b00f608 of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@2f17e30d of type java/util/ArrayList$Itr whose size is 32
I just allocated the object java.util.ArrayList$Itr@3e821657 of type java/util/ArrayList$Itr whose size is 32
I just allocated the object [Lio.netty.util.AsciiString;@1d131e1b of type io/netty/util/AsciiString whose size is 72
It's an array of size 14
I just allocated the object [Lio.netty.util.AsciiString;@71cf1b07 of type io/netty/util/AsciiString whose size is 96
It's an array of size 20
I just allocated the object key-0 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-1 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-2 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-3 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-4 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-5 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-6 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-7 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-8 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object key-9 of type io/netty/util/AsciiString whose size is 32
I just allocated the object glasgo of type io/netty/util/AsciiString whose size is 32
I just allocated the object io.grpc.netty.GrpcHttp2Headers@615091b8 of type io/grpc/netty/GrpcHttp2Headers whose size is 24
```