@carl-mastrangelo 2016-05-25
Sorry for the noise, but I have updated the benchmarks to be on top of the recent header changes I made.  It looks like it's in the realm of ~400ns for 10 client headers.